Rename Lookup interface; Startardize package name; Create lookup subpackage
