+++ middleware/src/main/java/br/ufrn/dimap/middleware/identification/ObjectId.java
@@ -1,5 +1,5 @@
 
-package identification;
+package br.ufrn.dimap.middleware.identification;
 
 import java.util.UUID;
